Cleanup and comment. Fix ItemBuilder processing functions even if there are none.

+++ src/main/java/me/hapyl/spigotutils/module/quest/objectives/TakeDamage.java
@@ -5,7 +5,7 @@ import me.hapyl.spigotutils.module.quest.QuestObjectiveType;
 import me.hapyl.spigotutils.module.util.Validate;
 
 public class TakeDamage extends QuestObjective {
-    public TakeDamage(long goalTotal) {
+    public TakeDamage(double goalTotal) {
         super(QuestObjectiveType.TAKE_DAMAGE, goalTotal, "Ouch!", String.format("Take %s damage from any source.", goalTotal));
     }
